+++ CHANGELOG.md
@@ -12,6 +12,8 @@
     [[#501]](https://github.com/extendr/extendr/pull/501)
 - Raw literals from Rust can be used for function and argument names. e.g. `fn r#type()` in Rust is converted to `type()` in R.
     [[#531]](https://github.com/extendr/extendr/pull/531)
+- Fix memory leaks on errors and panics
+    [[#555]](https://github.com/extendr/extendr/pull/555)
 
 ## 0.4.0